pyproject.toml
[tool.poetry]
name = "film2trello"
version = "0.0.0"
description = "Turns film links to Trello cards 🍿"
authors = ["Honza Javorek <mail@honzajavorek.cz>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/honzajavorek/film2trello/"

[tool.poetry.scripts]
film2trello = "film2trello:cli.main"

[tool.poetry.dependencies]
python = "3.11.*"
python-telegram-bot = "20.7"
click = "8.1.7"
httpx = {version = "0.25.*", extras = ["http2"]}
lxml = "5.0.0"
cssselect = "1.2.0"
pillow = "10.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.2"
ruff = "0.1.11"

[tool.pytest.ini_options]
python_files = "test_*.py"
testpaths = "."
norecursedirs = "legacy"
addopts = "--ff"

[tool.ruff]
exclude = ["legacy"]
target-version = "py311"

[tool.ruff.lint.isort]
combine-as-imports = true
# lines-after-imports = 2

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
